pyproject.toml
[project]
name = "live-comment-gemini"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "akaz", email = "82073147+akazdayo@users.noreply.github.com" }
]
dependencies = [
    "httpx>=0.27.0",
    "google-generativeai>=0.6.0",
    "pillow>=10.3.0",
    "opencv-python>=4.10.0.82",
    "numpy>=1.26.4",
    "pyautogui>=0.9.54",
    "vertexai>=1.49.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.rye.scripts]
start = {cmd = "python src/main.py", env-file = ".env"}

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/gemini_bot"]
